♻️ refactor : 자주 쓰이는 메서드 따로 분리 (createMapObj, createTotalCost)
Event클래스에 있던 메서드 (createMapObj, createTotalCost)를 utils로 이동.
