+++ src/utils/index.js
@@ -1,3 +1,7 @@
 export { default as OutputView } from './OutputView.js';
 export { default as InputView } from './InputView.js';
-export { default as createFoodObj } from './createFoodObj.js';
+export {
+  createFoodObj,
+  createMapObj,
+  createTotalCost,
+} from './createMethod.js';
